# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReleasePilot API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/demo/analyses": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_demo_analyses_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoAnalysisInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Recent",
        "operationId": "recent_api_v1_analyses_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisSummary"
                  },
                  "title": "Response Recent Api V1 Analyses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Start",
        "operationId": "start_api_v1_analyses_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Start Api V1 Analyses Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Report",
        "operationId": "report_api_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/findings": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Findings",
        "operationId": "findings_api_v1_analyses__analysis_id__findings_get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Finding"
                  },
                  "title": "Response Findings Api V1 Analyses  Analysis Id  Findings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/evidence": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Evidence",
        "operationId": "evidence_api_v1_analyses__analysis_id__evidence_get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Evidence"
                  },
                  "title": "Response Evidence Api V1 Analyses  Analysis Id  Evidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Documents",
        "operationId": "documents_api_v1_documents_get",
        "parameters": [
          {
            "name": "service_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Service Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Documents Api V1 Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Upload",
        "operationId": "upload_api_v1_documents_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_documents_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upload Api V1 Documents Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/search": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Retrieve",
        "operationId": "retrieve_api_v1_documents_search_get",
        "parameters": [
          {
            "name": "service_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Service Id"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2000,
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Retrieve Api V1 Documents Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "delete": {
        "tags": [
          "knowledge"
        ],
        "summary": "Remove",
        "operationId": "remove_api_v1_documents__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/reindex": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Reindex",
        "operationId": "reindex_api_v1_documents__document_id__reindex_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reindex Api V1 Documents  Document Id  Reindex Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisReport": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "release_candidate_id": {
            "type": "string",
            "format": "uuid",
            "title": "Release Candidate Id"
          },
          "service_name": {
            "type": "string",
            "title": "Service Name"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "source_mode": {
            "type": "string",
            "enum": [
              "demo",
              "github"
            ],
            "title": "Source Mode"
          },
          "state": {
            "type": "string",
            "enum": [
              "COMPLETED",
              "PARTIAL"
            ],
            "title": "State"
          },
          "base_sha": {
            "type": "string",
            "title": "Base Sha"
          },
          "head_sha": {
            "type": "string",
            "title": "Head Sha"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "release_candidate_id",
          "service_name",
          "scenario",
          "source_mode",
          "state",
          "base_sha",
          "head_sha",
          "created_at",
          "decision",
          "evidence"
        ],
        "title": "AnalysisReport"
      },
      "AnalysisSummary": {
        "properties": {
          "state": {
            "type": "string",
            "title": "State"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "service_name": {
            "type": "string",
            "title": "Service Name"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "source_mode": {
            "type": "string",
            "enum": [
              "demo",
              "github"
            ],
            "title": "Source Mode"
          },
          "recommendation": {
            "type": "string",
            "enum": [
              "GO",
              "CAUTION",
              "NO_GO"
            ],
            "title": "Recommendation"
          },
          "risk_score": {
            "type": "integer",
            "title": "Risk Score"
          },
          "confidence_score": {
            "type": "integer",
            "title": "Confidence Score"
          },
          "head_sha": {
            "type": "string",
            "title": "Head Sha"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "state",
          "id",
          "service_name",
          "scenario",
          "source_mode",
          "recommendation",
          "risk_score",
          "confidence_score",
          "head_sha",
          "created_at"
        ],
        "title": "AnalysisSummary"
      },
      "Body_upload_api_v1_documents_post": {
        "properties": {
          "service_id": {
            "type": "string",
            "format": "uuid",
            "title": "Service Id"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "service_id",
          "file"
        ],
        "title": "Body_upload_api_v1_documents_post"
      },
...
      "ConfidenceComponent": {
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "changes",
              "reviews",
              "ci",
              "runbooks"
            ],
            "title": "Category"
          },
          "weight": {
            "type": "integer",
            "title": "Weight"
          },
          "earned": {
            "type": "integer",
            "title": "Earned"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "category",
          "weight",
          "earned",
          "reason"
        ],
        "title": "ConfidenceComponent"
      },
...
      "Decision": {
        "properties": {
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "recommendation": {
            "type": "string",
            "enum": [
              "GO",
              "CAUTION",
              "NO_GO"
            ],
            "title": "Recommendation"
          },
          "risk_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score"
          },
          "confidence_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Confidence Score"
          },
          "category_scores": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "tests",
                "code_review",
                "complexity",
                "documentation"
              ]
            },
            "type": "object",
            "title": "Category Scores"
          },
          "category_caps": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "tests",
                "code_review",
                "complexity",
                "documentation"
              ]
            },
            "type": "object",
            "title": "Category Caps"
          },
          "confidence_components": {
            "items": {
              "$ref": "#/components/schemas/ConfidenceComponent"
            },
            "type": "array",
            "title": "Confidence Components"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "checklist": {
            "items": {
              "$ref": "#/components/schemas/ChecklistItem"
            },
            "type": "array",
            "title": "Checklist"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "policy_version",
          "recommendation",
          "risk_score",
          "confidence_score",
          "category_scores",
          "category_caps",
          "confidence_components",
          "summary",
          "findings",
          "checklist"
        ],
        "title": "Decision"
      },
      "DemoAnalysisInput": {
        "properties": {
          "service_id": {
            "type": "string",
            "format": "uuid",
            "title": "Service Id"
          },
          "scenario": {
            "type": "string",
            "enum": [
              "safe",
              "failed_ci",
              "missing_evidence"
            ],
            "title": "Scenario"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "service_id",
          "scenario"
        ],
        "title": "DemoAnalysisInput"
      },
      "Evidence": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "comparison",
              "review",
              "check",
              "runbook",
              "collection"
            ],
            "title": "Source Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "locator": {
            "type": "string",
            "title": "Locator"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source_type",
          "title",
          "excerpt",
          "locator",
          "captured_at"
        ],
        "title": "Evidence"
      },
      "Finding": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "rule": {
            "type": "string",
            "title": "Rule"
          },
          "category": {
            "type": "string",
            "enum": [
              "tests",
              "code_review",
              "complexity",
              "documentation"
            ],
            "title": "Category"
          },
          "kind": {
            "type": "string",
            "enum": [
              "blocker",
              "warning",
              "positive",
              "unknown"
            ],
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "remediation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remediation"
          },
          "points": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Points"
          },
          "evidence_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Evidence Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "rule",
          "category",
          "kind",
          "title",
          "explanation",
          "points",
          "evidence_ids"
        ],
        "title": "Finding"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StartInput": {
        "properties": {
          "service_id": {
            "type": "string",
            "format": "uuid",
            "title": "Service Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "github"
            ],
            "title": "Mode",
            "default": "demo"
          },
          "scenario": {
            "type": "string",
            "enum": [
              "safe",
              "failed_ci",
              "missing_evidence"
            ],
            "title": "Scenario",
            "default": "failed_ci"
          },
          "base_ref": {
            "type": "string",
            "maxLength": 200,
            "title": "Base Ref",
            "default": ""
          },
          "head_ref": {
            "type": "string",
            "maxLength": 200,
            "title": "Head Ref",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "service_id"
        ],
        "title": "StartInput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}